requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
